Rebase and refactor after resource proxy check in
