--- src/components/package/_fields/proxy-select/index.js
@@ -1,1 +1,0 @@
-export { default } from './proxy-select-field';
