--- src/components/package/_fields/proxy-select/proxy-select-field.css
@@ -1,3 +1,0 @@
-.proxy-select-field {
-  max-width: 50em;
-}
