+++ src/components/package/edit-managed/managed-package-edit.js
@@ -20,7 +20,7 @@ import NavigationModal from '../../navigation-modal';
 import Toaster from '../../toaster';
 import PaneHeaderButton from '../../pane-header-button';
 import SelectionStatus from '../selection-status';
-import ProxySelectField from '../_fields/proxy-select';
+import ProxySelectField from '../../proxy-select';
 import styles from './managed-package-edit.css';
 
 class ManagedPackageEdit extends Component {
@@ -317,10 +317,12 @@ class ManagedPackageEdit extends Component {
                         )}
                       </div>
                       {(proxyTypes.request.isResolved && provider.data.isLoaded) ? (
-                        <ProxySelectField
-                          proxyTypes={proxyTypes}
-                          inheritedProxy={provider.proxy}
-                        />
+                        <div data-test-eholdings-package-proxy-select-field>
+                          <ProxySelectField
+                            proxyTypes={proxyTypes}
+                            inheritedProxyId={provider.proxy.id}
+                          />
+                        </div>
                       ) : (
                         <Icon icon="spinner-ellipsis" />
                       )}
